+++ common/src/main/java/mca/Config.java
@@ -66,7 +66,7 @@ public final class Config implements Serializable {
     public float maleVillagerWidthFactor = 1.0f;
     public float femaleVillagerWidthFactor = 0.95f;
     public boolean showNameTags = true;
-    public boolean useMcaVoices = false;
+    public boolean useMCAVoices = true;
     public boolean useVanillaVoices = false;
     public float interactionChanceFatigue = 1.0f;
     public int interactionFatigueCooldown = 4800;