+++ fabric/src/main/java/mca/fabric/MCAFabricClient.java
@@ -21,6 +21,7 @@ import mca.item.SirbenBabyItem;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.BlockEntityRendererRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
@@ -73,6 +74,10 @@ public final class MCAFabricClient extends ClientProxyAbstractImpl implements Cl
         );
         ModelPredicateProviderRegistry.register(ItemsMCA.SIRBEN_BABY_GIRL.get(), new Identifier("invalidated"), (stack, world, entity, i) ->
                 SirbenBabyItem.hasBeenInvalidated(stack) ? 1 : 0
+        );
+
+        ClientPlayConnectionEvents.JOIN.register((handler, sender, server) ->
+                MCAClient.onLogin()
         );
 
         BlockRenderLayerMap.INSTANCE.putBlock(BlocksMCA.INFERNAL_FLAME.get(), RenderLayer.getCutout());